# openapi.yaml
openapi: 3.0.3
info:
  title: ikctl API
  description: API REST para gestión de instalaciones remotas de aplicaciones vía SSH
  version: 1.0.0
  contact:
    name: ikctl Team
    email: support@ikctl.com
servers:
  - url: http://localhost:8000
    description: Desarrollo local
  - url: https://api.ikctl.com
    description: Producción
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/operations:
    post:
      tags: [operations]
      summary: Lanzar operación (RF-14)
      description: >
        Crea la operación con estado pending y la ejecuta de forma asíncrona.
        El kit debe tener sync_status synced y el servidor debe estar active.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationLaunch'
      responses:
        '201':
          description: Operación creada y encolada
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Operation'
                  - $ref: '#/components/schemas/BatchOperationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    get:
      tags: [operations]
      summary: Listar operaciones del usuario (RF-16)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: server_id
          in: query
          schema:
            type: string
        - name: kit_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, failed, cancelled, cancelled_unsafe]
      responses:
        '200':
          description: Lista de operaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOperations'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/operations/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [operations]
      summary: Consultar estado de operación (RF-15)
      description: >
        output solo se devuelve si debug_level es errors o full.
        Con errors devuelve solo stderr; con full devuelve stdout + stderr.
        Soporta polling mientras status es in_progress.
      responses:
        '200':
          description: Estado de la operación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/operations/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      tags: [operations]
      summary: Cancelar operación (RF-17)
      description: >
        pending → cancelled (limpio).
        in_progress → cancelled_unsafe (servidor puede quedar en estado parcial).
      responses:
        '200':
          description: Operación cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: La operación ya está en estado terminal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/operations/{id}/restore:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      tags: [operations]
      summary: Restaurar backup pre-operación (RF-18)
      description: >
        Disponible para operaciones failed o cancelled_unsafe con backup[] declarado en el kit.
        Ejecuta cp {path}.bak.ikctl {path} para cada fichero del backup en el servidor.
      responses:
        '200':
          description: Backup restaurado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Backup restaurado correctamente"
                  restored_files:
                    type: array
                    items:
                      type: string
                    example: ["/etc/nginx/nginx.conf"]
        '400':
          description: No hay backup disponible para esta operación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/operations/{id}/retry:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      tags: [operations]
      summary: Reintentar operación fallida (RF-19)
      description: >
        Disponible para operaciones en estado failed o cancelled_unsafe.
        Crea una nueva operación con los mismos parámetros conservando la original en el historial.
      responses:
        '202':
          description: Nueva operación creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: La operación no está en estado retriable (failed o cancelled_unsafe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── PIPELINES ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid_credentials"
        message:
          type: string
          example: "Credenciales inválidas"
        details:
          type: object
          additionalProperties: true

    # ── Credentials ───────────────────────────────────────────────────────────
...
    Operation:
      type: object
      properties:
        operation_id:
          type: string
          example: "op_1a2b3c4d"
        user_id:
          type: string
        server_id:
          type: string
          example: "srv_1a2b3c4d"
        kit_id:
          type: string
          example: "kit_1a2b3c4d"
        values:
          type: object
          additionalProperties: true
          description: "Valores efectivos usados en la ejecución (defaults del kit + overrides del request)"
        sudo:
          type: boolean
          example: false
        debug_level:
          type: string
          enum: [none, errors, full]
          example: none
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled, cancelled_unsafe]
          example: completed
        output:
          type: string
          nullable: true
          description: "Solo se devuelve si debug_level es errors o full"
        backup_files:
          type: array
          items:
            type: string
          description: "Rutas de archivos de backup creados durante la operación"
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OperationLaunch:
      type: object
      required: [kit_id]
      description: >
        Exactamente uno de `server_id` o `group_id` debe estar presente.
      properties:
        server_id:
          type: string
          nullable: true
          example: "srv_1a2b3c4d"
        group_id:
          type: string
          nullable: true
          example: "grp_1a2b3c4d"
          description: "Si se especifica, lanza la operación en todos los servidores del grupo (batch)"
        kit_id:
          type: string
          example: "kit_1a2b3c4d"
        values:
          type: object
          additionalProperties: true
          default: {}
          description: "Sobreescriben los valores por defecto del kit"
        sudo:
          type: boolean
          default: false
        debug_level:
          type: string
          enum: [none, errors, full]
          description: "Si no se especifica, hereda el debug_level del ikctl.yaml del kit"
        timeout_seconds:
          type: integer
          default: 1800
          example: 600
    PaginatedOperations:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Operation'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
    BatchOperationResponse:
      type: object
      description: "Respuesta cuando se lanza una operación en batch sobre un grupo de servidores"
      properties:
        operations:
          type: array
          items:
            $ref: '#/components/schemas/Operation'

    # ── Pipelines ─────────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No autenticado o token inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Token inválido o expirado"
    Forbidden:
      description: Acceso denegado — el recurso no pertenece al usuario autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "No tienes permisos para acceder a este recurso"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Recurso no encontrado"
...
    UnprocessableEntity:
      description: Datos válidos en formato pero inválidos en lógica de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unprocessable_entity"
            message: "El kit debe tener sync_status synced para lanzar una operación"